ref: Split post process forwarder from the eventstream
It doesn't make sense for the post process forwarder to be part of the `eventstream` module.
This unnecessarily couples the ability to do post processing an dataset, to the use of the
`Event`/`GroupEvent` model (which is complex and probably overused - why is everything an "event").

With this change post processing can be easily added on any dataset, regardless of whether an event
actually uses the `Event`/`GroupEvent` model.

There are intentionally no functional changes or refactorings here to make review easier - just the
minimal changes needed to move files around and split the eventstream module. Post processing for
errors, transactions and issue occurrences still works in exactly the same way as before.
