+++ mypy.ini
@@ -98,6 +98,7 @@ files = fixtures/mypy-stubs,
         src/sentry/ownership/grammar.py,
         src/sentry/pipeline/,
         src/sentry/plugins/base/view.py,
+        src/sentry/post_process_forwarder/,
         src/sentry/processing/realtime_metrics/,
         src/sentry/profiles/,
         src/sentry/ratelimits/,
@@ -184,9 +185,8 @@ files = fixtures/mypy-stubs,
         src/sentry/web/decorators.py,
         src/sentry/web/frontend/base.py,
         src/sentry/web/helpers.py,
-        tests/sentry/eventstream/kafka/test_consumer_strategy.py,
-        tests/sentry/eventstream/kafka/test_synchronized.py,
         tests/sentry/lang/native/test_appconnect.py,
+        tests/sentry/post_process_forwarder,
         tests/sentry/processing/realtime_metrics/,
         tests/sentry/issues/,
         tests/sentry/tasks/test_low_priority_symbolication.py,
