+++ src/sentry/post_process_forwarder/__init__.py
@@ -1,0 +1,3 @@
+from .post_process_forwarder import PostProcessForwarder, PostProcessForwarderType
+
+__all__ = ["PostProcessForwarder", "PostProcessForwarderType"]
